Fix remarks on review of #8008
Rename nieuwsblad to more generic publisher 'Mediahuis' and add support for:
    - Het Nieuwsblad
    - De Standaard
    - Gazet van Antwerpen
    - Het Belang van Limburg
